\name{ais}
\alias{AIS}
\alias{ais}
\title{Australian athletes data}
\non_function{}
\description{
Data on 102 male and 100 female athletes collected at the Australian
Institute of Sport, courtesy of Richard Telford and Ross Cunningham.
}
\format{This data frame contains the following columns:
 \describe{
    \item{Sex}{
 (0 = male or 1 = female)
    }
    \item{Ht}{
 height (cm)
    }
    \item{Wt}{
 weight (kg)
    }
    \item{LBM}{
 lean body mass
    }
    \item{RCC}{
 red cell count
    }
    \item{WCC}{
 white cell count
    }
    \item{Hc}{
 Hematocrit
    }
    \item{Hg}{
 Hemoglobin
    }
 }
}
\details{}
\source{R. D. Cook and S. Weisberg (1999). \emph{Applied
Statistics Including Computing and Graphics}.  New York:  Wiley.}
\keyword{datasets}
