% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{dbr_estimator}
\alias{dbr_estimator}
\title{Makes DBR estimators for group-level data}
\usage{
dbr_estimator(data, models, randomization, hajek = FALSE, ...)
}
\description{
Makes DBR estimators for group-level data
}

