% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{dr_estimators}
\alias{dr_estimators}
\title{Makes IPW, OTC, and DBR estimators for group-level data}
\usage{
dr_estimators(data, t_model, o_model, randomization)
}
\description{
Makes IPW, OTC, and DBR estimators for group-level data
}

