% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{evaluate_df_function}
\alias{evaluate_df_function}
\title{#'-----------------------------------------------------------------------------#
#' Fills a matrix with zeros
#' @export
#'-----------------------------------------------------------------------------#}
\usage{
evaluate_df_function(f, ...)
}
\description{
rbind_fill_zero <- function(this_list)
{
  if(!is.list(this_list)){
    this_list <- list(this_list)
  }
  p <- max(unlist(lapply(this_list, ncol)))
  out <- lapply(this_list, function(x){
    xp <- ncol(x)
    if(xp < p ){
      cbind(x,matrix(0, nrow = nrow(x), ncol = p - xp))
    } else {
      x
    }
  })
  do.call('rbind', out)
}
Evaluate a function on a dataframe
}
\details{
For use in dplyr
}
