% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extract_model_info}
\alias{extract_model_info}
\title{Extract relevant information from list of models}
\usage{
extract_model_info(models, data, estimator_type, regression_type = "none")
}
\description{
Extract relevant information from list of models
}

