% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eefuns.R
\name{generic_eefun}
\alias{generic_eefun}
\title{Estimating equations for IPW, OTC, DBR, and WLS_DBR estimators}
\usage{
generic_eefun(data, models, randomization = 1, estimator_type,
  regression_type = "none", hajek = FALSE)
}
\description{
Estimating equations for IPW, OTC, DBR, and WLS_DBR estimators
}

