% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_design_frame}
\alias{get_design_frame}
\title{Get a matrix of fixed effects from a model object}
\usage{
get_design_frame(rhs_formula, data)
}
\description{
Get a matrix of fixed effects from a model object
}

