% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{ipw_estimator}
\alias{ipw_estimator}
\title{Makes IPW estimator for group-level data}
\usage{
ipw_estimator(data, models, randomization, hajek = FALSE, ...)
}
\description{
Makes IPW estimator for group-level data
}

