% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eefuns.R
\name{make_eefun_wls}
\alias{make_eefun_wls}
\title{Estimating equations for WLS scores}
\usage{
make_eefun_wls(t_model, wls_model, data, a, randomization = 1)
}
\description{
Estimating equations for WLS scores
}

