\name{mussels}
\alias{mussels}
\title{Mussels' muscles data}
\non_function{}
\description{
Data were furnished by Mike Camden, Wellington Polytechnic, Wellington,
New Zealand.  Horse mussels, (Atrinia), were sampled from the Marlborough
Sounds.  The response is the mussels' Muscle Mass.
}
\format{This data frame contains the following columns:
 \describe{
    \item{H}{
 Shell height in mm
    }
    \item{L}{
 Shell length in mm
    }
    \item{M}{
 Muscle mass in g
    }
    \item{S}{
 Shell mass in g
    }
    \item{W}{
 Shell width in mm
    }
 }
}

\source{R. D. Cook and S. Weisberg (1999). \emph{Applied
Statistics Including Computing and Graphics}.  New York:  Wiley.}
\keyword{datasets}
