% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{otc_estimator}
\alias{otc_estimator}
\title{Makes OTC (outcome based) estimator for group-level data}
\usage{
otc_estimator(data, models, randomization, ...)
}
\description{
Makes OTC (outcome based) estimator for group-level data
}
