% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{pcov_dbr_estimator}
\alias{pcov_dbr_estimator}
\title{Makes regression with propensity (DBR) estimators for group-level data
(i.e.) regression based}
\usage{
pcov_dbr_estimator(data, models, randomization, ...)
}
\description{
Makes regression with propensity (DBR) estimators for group-level data
(i.e.) regression based
}

