% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{reg_dbr_estimator}
\alias{reg_dbr_estimator}
\title{Makes WLS (DBR) estimators for group-level data
(i.e.) regression based}
\usage{
reg_dbr_estimator(data, models, randomization, regression_type, ...)
}
\description{
Makes WLS (DBR) estimators for group-level data
(i.e.) regression based
}

