% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eefuns.R
\name{wls_dbr_estimator_estFUN}
\alias{wls_dbr_estimator_estFUN}
\title{Estimator used in wls estFUN}
\usage{
wls_dbr_estimator_estFUN(data, models, randomization, regression_type = "wls",
  ...)
}
\description{
Estimator used in wls estFUN
}
