% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eefuns.R
\name{wls_score_maker}
\alias{wls_score_maker}
\title{Estimating equations for WLS scores}
\usage{
wls_score_maker(theta, X, Y, phi, ipw, A, a, family_link, invlnk)
}
\description{
Estimating equations for WLS scores
}
