% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eefuns.R
\name{generic_eefun}
\alias{generic_eefun}
\title{Estimating equations for IPW, OTC, and DBR estimators}
\usage{
generic_eefun(data, models, randomization, estimator_type,
  estimator_label = NULL, hajek = FALSE)
}
\description{
Estimating equations for IPW, OTC, and DBR estimators
}

