% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eefuns.R
\name{generic_eefun}
\alias{generic_eefun}
\title{Estimating equations for IPW, OTC, and DBR estimators}
\usage{
generic_eefun(data, t_model, o_model, randomization, estimatorFUN)
}
\description{
Estimating equations for IPW, OTC, and DBR estimators
}

