% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{wls_dbr_estimator}
\alias{wls_dbr_estimator}
\title{Makes WLS (DBR based) estimator for group-level data}
\usage{
wls_dbr_estimator(data, models, randomization, ...)
}
\description{
Makes WLS (DBR based) estimator for group-level data
}

